Refactor to make fields a subpackage.
