+++ binobj/fields/__init__.py
@@ -1,0 +1,6 @@
+"""Definitions of all field types that can be used in structures."""
+
+from .base import *
+from .containers import *
+from .numeric import *
+from .stringlike import *
